Move CUDA files into Platforms/CUDA.
