+++ src/Platforms/CMakeLists.txt
@@ -23,7 +23,7 @@ SET(DEVICE_SRCS
 ADD_LIBRARY(platform_host ${HOST_SRCS})
 
 ADD_LIBRARY(platform_device ${DEVICE_SRCS})
-TARGET_INCLUDE_DIRECTORIES(platform_device INTERFACE "${CMAKE_CURRENT_SOURCE_DIR}")
+TARGET_INCLUDE_DIRECTORIES(platform_device PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}")
 TARGET_LINK_LIBRARIES(platform_device PUBLIC message)
 
 SUBDIRS(OpenMP)
