+++ src/Platforms/CUDA/CMakeLists.txt
@@ -15,3 +15,8 @@ SET(CUDA_SRCS
 )
 
 CUDA_ADD_LIBRARY(platform_cuda ${CUDA_SRCS})
+TARGET_INCLUDE_DIRECTORIES(platform_cuda INTERFACE "${CMAKE_CURRENT_SOURCE_DIR}")
+
+IF (BUILD_UNIT_TESTS)
+  SUBDIRS(tests)
+ENDIF()
