+++ src/Platforms/CUDA/CUDAallocator.hpp
@@ -23,7 +23,7 @@
 #include <stdexcept>
 #include <cuda_runtime_api.h>
 #include "config.h"
-#include "CUDA/cudaError.h"
+#include "cudaError.h"
 #include "simd/allocator_traits.hpp"
 
 namespace qmcplusplus
