+++ src/Platforms/CUDA/tests/CMakeLists.txt
@@ -10,18 +10,13 @@
 #//////////////////////////////////////////////////////////////////////////////////////
 
 
-IF(ENABLE_CUDA)
-
 SET(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${QMCPACK_UNIT_TEST_DIR})
 
 SET(SRC_DIR CUDA)
 SET(UTEST_EXE test_${SRC_DIR})
 SET(UTEST_NAME deterministic-unit_test_${SRC_DIR})
 
-
 ADD_EXECUTABLE(${UTEST_EXE} test_CUDAallocator.cpp)
-TARGET_LINK_LIBRARIES(${UTEST_EXE} catch_main qmcutil)
+TARGET_LINK_LIBRARIES(${UTEST_EXE} platform_cuda catch_main)
 
 ADD_UNIT_TEST(${UTEST_NAME} "${QMCPACK_UNIT_TEST_DIR}/${UTEST_EXE}")
-
-ENDIF(ENABLE_CUDA)
